refactor: split summary report facade
